refactor: :recycle: Splitted dialogs in dedicated components
+ loading in store
+ form validation on new components
Still need to do agenda
