+++ components/AccountList.vue
@@ -1,65 +1,6 @@
 <template>
   <div>
-    <v-dialog v-model="dialog" width="500">
-      <v-card>
-        <v-form v-model="valid" @submit="createAccount">
-          <v-toolbar elevation="0" dense>
-            <v-toolbar-title>
-              {{ account.id ? 'Editer' : 'Créer' }} un compte
-            </v-toolbar-title>
-            <v-spacer></v-spacer>
-            <v-btn icon color="grey" small plain @click="dialog = false">
-              <v-icon>mdi-close</v-icon>
-            </v-btn>
-          </v-toolbar>
-
-          <v-card-text>
-            <v-container>
-              <v-row>
-                <v-col>
-                  <v-text-field
-                    v-model="account.name"
-                    label="Nom du compte"
-                    required
-                    :dense="$vuetify.breakpoint.smAndDown"
-                  >
-                  </v-text-field>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col>
-                  <v-text-field
-                    v-model="account.balance"
-                    label="Solde du compte"
-                    type="number"
-                    prefix="€"
-                    :dense="$vuetify.breakpoint.smAndDown"
-                  >
-                  </v-text-field>
-                </v-col>
-              </v-row>
-            </v-container>
-          </v-card-text>
-
-          <v-divider></v-divider>
-
-          <v-card-actions>
-            <v-spacer></v-spacer>
-            <v-btn color="error" text @click="dialog = false"> Annuler </v-btn>
-            <v-btn
-              color="success"
-              :loading="loading"
-              :disabled="!valid"
-              text
-              type="submit"
-              :dense="$vuetify.breakpoint.smAndDown"
-            >
-              Valider
-            </v-btn>
-          </v-card-actions>
-        </v-form>
-      </v-card>
-    </v-dialog>
+    <AccountEditionDialog v-model="account" />
     <v-card>
       <v-card-title>
         <span class="font-weight-light">Comptes</span>
@@ -76,7 +17,7 @@
             icon
             color="success"
             :small="$vuetify.breakpoint.smAndDown"
-            @click="showNew"
+            @click="account = undefined"
           >
             <v-icon>mdi-plus</v-icon>
           </v-btn>
@@ -129,23 +70,21 @@
 
 <script lang="ts">
 import Vue from 'vue'
+import { mapGetters } from 'vuex'
+import AccountEditionDialog from './dialogs/AccountEditionDialog.vue'
 import type { Account, InputAccount } from '~/ts/types'
 import { toLS } from '~/ts/format'
 
 export default Vue.extend({
+  components: { AccountEditionDialog },
   data: () => ({
     selectedItem: undefined,
-    initAccount: {
-      id: undefined as string | undefined,
-      name: '',
-      balance: '0',
-    },
-    account: {} as InputAccount,
-    valid: true,
-    dialog: false,
-    loading: false,
+    account: null as InputAccount | null,
   }),
   computed: {
+    ...mapGetters({
+      loading: 'account/getLoadingState',
+    }),
     /**
      * The sub-accounts of the user
      */
@@ -174,28 +113,6 @@ export default Vue.extend({
   methods: {
     toLS,
     /**
-     * Create sub account
-     */
-    async createAccount(e: Event) {
-      e.preventDefault()
-      if (this.valid) {
-        this.loading = true
-        try {
-          if (this.account.id) {
-            await this.$store.dispatch('account/editAccount', this.account)
-            this.$toast.global.success('Compte edité')
-          } else {
-            await this.$store.dispatch('account/createAccount', this.account)
-            this.$toast.global.success('Compte créé')
-          }
-        } catch (e) {
-          this.$toast.global.error((e as Error).message)
-        }
-        this.dialog = false
-        this.loading = false
-      }
-    },
-    /**
      * Delete sub account
      * @param i The index in `this.accounts`
      */
@@ -216,7 +133,6 @@ export default Vue.extend({
       })
 
       if (res) {
-        this.loading = true
         try {
           await this.$store.dispatch(
             'account/deleteAccount',
@@ -226,26 +142,15 @@ export default Vue.extend({
         } catch (e) {
           this.$toast.global.error((e as Error).message)
         }
-        this.loading = false
       }
-    },
-    /**
-     * Show popup for adding a new sub account
-     */
-    showNew() {
-      // this.valid = false
-      this.account = { ...this.initAccount }
-      this.dialog = true
     },
     /**
      * Show popup for editing a sub account
      * @param i The index in `this.accounts`
      */
     showEdit(i: number) {
-      this.valid = true
       const acc = this.accounts[i]
       this.account = { ...acc, balance: acc.balance.toFixed(2), id: acc.id }
-      this.dialog = true
     },
   },
 })
