+++ store/operations/actions.ts
@@ -22,37 +22,43 @@ const actions: ActionTree<OperationState, RootState> = {
     { name, amount, category, modifier, date }: InputOperation
   ) {
     commit('SET_LOADING', true)
-    const uid = (rootGetters['auth/getUser'] as User | null)?.uid
-    if (!uid) {
-      throw new Error('Vous devez être connecté pour effectuer cette action')
+    try {
+      const uid = (rootGetters['auth/getUser'] as User | null)?.uid
+      if (!uid) {
+        throw new Error('Vous devez être connecté pour effectuer cette action')
+      }
+
+      const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
+      if (!aid) {
+        throw new Error('Un compte doit être séléctionné')
+      }
+
+      const amnt = parseFloat(amount) * modifier
+      const ref = this.$fire.firestore
+        .collection('users')
+        .doc(uid)
+        .collection('accounts')
+        .doc(aid)
+
+      // Get category reference
+      let cref = null
+      if (category && typeof category === 'string')
+        cref = ref.collection('categories').doc(category)
+
+      const ope = await ref.collection('operations').add({
+        name,
+        amount: amnt,
+        category: cref,
+        date: date && this.$fireModule.firestore.Timestamp.fromDate(date),
+        createdAt: this.$fireModule.firestore.FieldValue.serverTimestamp(),
+      } as Operation & { createdAt: firebase.firestore.FieldValue })
+
+      commit('SET_LOADING', false)
+      return ope
+    } catch (error) {
+      commit('SET_LOADING', false)
+      throw error
     }
-
-    const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
-    if (!aid) {
-      throw new Error('Un compte doit être séléctionné')
-    }
-
-    const amnt = parseFloat(amount) * modifier
-    const ref = this.$fire.firestore
-      .collection('users')
-      .doc(uid)
-      .collection('accounts')
-      .doc(aid)
-
-    // Get category reference
-    let cref = null
-    if (category && typeof category === 'string')
-      cref = ref.collection('categories').doc(category)
-
-    const ope = await ref.collection('operations').add({
-      name,
-      amount: amnt,
-      category: cref,
-      date: date && this.$fireModule.firestore.Timestamp.fromDate(date),
-      createdAt: this.$fireModule.firestore.FieldValue.serverTimestamp(),
-    } as Operation & { createdAt: firebase.firestore.FieldValue })
-    commit('SET_LOADING', false)
-    return ope
   },
   /**
    * Edit an operation for the auther user in the selected account
@@ -66,42 +72,46 @@ const actions: ActionTree<OperationState, RootState> = {
     { id, name, amount, category, modifier, date }: InputOperation
   ) {
     commit('SET_LOADING', true)
-    const uid = (rootGetters['auth/getUser'] as User | null)?.uid
-    if (!uid) {
-      throw new Error('Vous devez être connecté pour effectuer cette action')
+    try {
+      const uid = (rootGetters['auth/getUser'] as User | null)?.uid
+      if (!uid) {
+        throw new Error('Vous devez être connecté pour effectuer cette action')
+      }
+
+      const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
+      if (!aid) {
+        throw new Error('Un compte doit être séléctionné')
+      }
+
+      // Parse amount & category
+      const amnt = parseFloat(amount) * modifier
+      if (category && typeof category !== 'string') {
+        category = category.id
+      }
+
+      const ref = this.$fire.firestore
+        .collection('users')
+        .doc(uid)
+        .collection('accounts')
+        .doc(aid)
+      const cref = ref.collection('categories')
+
+      await ref
+        .collection('operations')
+        .doc(id)
+        .update({
+          name,
+          amount: amnt,
+          category: category ? cref.doc(category) : null,
+          date: date && this.$fireModule.firestore.Timestamp.fromDate(date),
+          updatedAt: this.$fireModule.firestore.FieldValue.serverTimestamp(),
+        } as Operation & { updatedAt: firebase.firestore.FieldValue })
+
+      commit('SET_LOADING', false)
+    } catch (error) {
+      commit('SET_LOADING', false)
+      throw error
     }
-
-    const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
-    if (!aid) {
-      throw new Error('Un compte doit être séléctionné')
-    }
-
-    // Parse amount & category
-    const amnt = parseFloat(amount) * modifier
-    if (category && typeof category !== 'string') {
-      category = category.id
-    }
-
-    const ref = this.$fire.firestore
-      .collection('users')
-      .doc(uid)
-      .collection('accounts')
-      .doc(aid)
-    const cref = ref.collection('categories')
-
-    const ope = await ref
-      .collection('operations')
-      .doc(id)
-      .update({
-        name,
-        amount: amnt,
-        category: category ? cref.doc(category) : null,
-        date: date && this.$fireModule.firestore.Timestamp.fromDate(date),
-        updatedAt: this.$fireModule.firestore.FieldValue.serverTimestamp(),
-      } as Operation & { updatedAt: firebase.firestore.FieldValue })
-
-    commit('SET_LOADING', false)
-    return ope
   },
   /**
    * Delete an operation for the auther user in the selected account
@@ -110,27 +120,33 @@ const actions: ActionTree<OperationState, RootState> = {
    * @param id The id of the operation
    * @returns The promise of deletion
    */
-  deleteOperation({ rootGetters }, id) {
-    // const amnt = parseFloat(amount)
-    const uid = (rootGetters['auth/getUser'] as User | null)?.uid
-    if (!uid) {
-      throw new Error('Vous devez être connecté pour effectuer cette action')
+  async deleteOperation({ rootGetters, commit }, id) {
+    commit('SET_LOADING', true)
+    try {
+      const uid = (rootGetters['auth/getUser'] as User | null)?.uid
+      if (!uid) {
+        throw new Error('Vous devez être connecté pour effectuer cette action')
+      }
+
+      const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
+      if (!aid) {
+        throw new Error('Un compte doit être séléctionné')
+      }
+
+      const ref = this.$fire.firestore
+        .collection('users')
+        .doc(uid)
+        .collection('accounts')
+        .doc(aid)
+        .collection('operations')
+        .doc(id)
+
+      await ref.delete()
+      commit('SET_LOADING', false)
+    } catch (error) {
+      commit('SET_LOADING', false)
+      throw error
     }
-
-    const aid = (rootGetters['account/getCurrent'] as Account | null)?.id
-    if (!aid) {
-      throw new Error('Un compte doit être séléctionné')
-    }
-
-    const ref = this.$fire.firestore
-      .collection('users')
-      .doc(uid)
-      .collection('accounts')
-      .doc(aid)
-      .collection('operations')
-      .doc(id)
-
-    return ref.delete()
   },
   /**
    * Bind user's operations for the auther user in the selected account
@@ -181,7 +197,7 @@ const actions: ActionTree<OperationState, RootState> = {
       .orderBy('date', 'desc')
       .orderBy('createdAt', 'desc')
 
-    await bindFirestoreRef('data', docref, {
+    const bind = await bindFirestoreRef('data', docref, {
       serialize: (doc) => {
         const data = doc.data()
         Object.defineProperty(data, 'id', { value: doc.id })
@@ -189,6 +205,7 @@ const actions: ActionTree<OperationState, RootState> = {
       },
     })
     commit('SET_LOADING', false)
+    return bind
   }),
 }
